docs/spec/hooks/examples/python/pyproject.toml
[project]
name = "hooks-interchange"
version = "0.1.0"
description = "Reference implementation for the Hook Interchange Format Specification (Core conformance)"
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
